Change BLE packet GSR payload type from int to float

+++ app/src/main/java/pl/edu/pjwstk/engineeringthesis/bluetooth/Packet.java
@@ -6,14 +6,14 @@ import java.util.List;
 
 public class Packet {
     public final List<Float> temps;
-    public final List<Integer> gsr;
+    public final List<Float> gsr;
     public final List<Float> hearthRate;
     public final List<Float> spo2;
 
     public final long epoch;
     public final JSONObject raw;
 
-    public Packet(ArrayList<Float> temps, ArrayList<Integer> gsr, ArrayList<Float> hearthRate, ArrayList<Float> spo2, long epoch, JSONObject raw) {
+    public Packet(ArrayList<Float> temps, ArrayList<Float> gsr, ArrayList<Float> hearthRate, ArrayList<Float> spo2, long epoch, JSONObject raw) {
         this.temps = temps;
         this.gsr = gsr;
         this.hearthRate = hearthRate;
